checkpoint commit
Switched from overrideDerivation to overrideAttrs because
overrideDerivation was not working on Russell's new PDF derivation --
see https://github.com/BlockstreamResearch/simplicity/pull/145

Russell explained on Rocket that overrideDerivation is "old school"
and soft-deprecated and that overrideAttrs is to be prefered. See
https://nixos.org/manual/nixpkgs/stable/#sec-pkg-overrideDerivation
(It references "the overrideAttrs documentation". To find this, scroll
up.)