checkpoint: have new matrix function which can do functions, rust-bitcoin using it
Goal is to get rid of `argsMatrices` and just have one matrix which
defines all the matrix arguments, possibly in terms of each other (which
is done using Nix functions and some clever ordering logic). We have
some further optimization to do -- I believe this is very slow mainly
due to re-evaluation of `gitCommits` which is an IFD which calls git a
bunch of times, so the next step is to try to convert that from a list
to a map, which will have better locality properties.

Update rust-bitcoin.check-pr.nix to use this. Probably other check-prs
are broken. But the hope is that this will eventually be so easy to use
that it'll be easy to just re-do everything else.

One new feature is the ability to get the list of workspaces from the
TOML file. This has allowed me to correct my list of workspaces (was
missing a couple). And also sanity check feature lists a bit (was also
way out of date). Also added fuzzing to the rust-bitcoin repo.