checkpoint commit: upgrade rust-jsonrpc, update crate2nix for recent nixpkgs
After upgrading my nixpkgs channel I started getting mysterious breakage
where crate2nix would show "Did not find all crates.io hashes in
Cargo.lock" and be unable to produce the generated/called nixfiles.

It appears the culprit is somewhere in https://github.com/NixOS/nixpkgs/pull/303527
and it was fixed in 640584db14761545480c0d29b80c3ab34a3236bd of
crate2nix (part of #341). We previously had been pinned to an old
version of crate2nix due to incompatibility with cargo 1.48 but
fortunately the rust-bitcoin ecosystem has since moved on to 1.56.1 so
we are ok to upgrade.