utils: move calling of the IFD out of the memo function
We have a memoization function whose goal is to avoid redundant calls to
`callPackage` on an IFD. We create a "memo key" where we stringify all
the inputs both to the IFD generation and to its calling, and then key
the resulting derivation on that.

The idea is that there are 3 ways that the matrix parameters eventually
lead to code being written:

1 Some (src, lockfile) parameterize the IFD itself, i.e. the internally
  generated Cargo.nix is templated on these. We call the resulting
  derivation the "generated derivation".
2 Some (rustc, release mode) are given to Cargo.nix as arguments via
  pkgs.callPackage. The output of this (which is expensive to produce)
  we call the "called derivation".
3 Some (runDocs, runClippy, etc) are used in overrides in outputs of the
  called derivation. Also expensive to produce.

Our existing logic memoizes steps 1 and 2. This commit memoizes only
step 1. Since each generated derivation produces many called
derivations, this reduces the size of the memo table, and is able to
reuse more generated Cargo.nixes (at the expense of evaluating each one
more times).

Since the evaluations in step 1 are the ones that call `crate2nix
generate`, while those in step 2 do a lot of nix-internal work, it's not
intuitively obvious what the effect on performance should be. In fact,
it appears to have almost no effect on the speed of instantiation, but
it _does_ make the instantiation output cleaner and easier to read
(since there are fewer calls to `crate2nix generate` and each one is
shorter and simpler, and there is much more time between them).

So we consider this a win even though it doesn't meaningfully speed
things up.

To meaningfully speed things up, likely we would need to move more steps
out of Nix entirely into IFDs.