checkpoint commit: fix nightly check, speed up linkfarm a bit, have public API check almost working
There was a bug in our matching on rust version which caused us to never
recognize a rustc as "nightly", meaning that lots of checks weren't
being done. Fix that.

Also, do some benchmarking work and find that `linkFarm` does a really
inefficient foldl thing if you give it a list, with the goal of handling
duplicate keys (which we won't have) in an elegant way. So call
listToAttrs before calling linkFarm, which causes it to skip this check.

Also, continue work on getting the cargo-public-api stuff to work. Seems
like we are close.

Meanwhile I think there is some more room for optimization by doing two
layers of memoization: one for the generated cargo nix (done now) and
one for the called cargo nix (not yet done).