checkpoint commit; support multiple honggfuzz versions; run doctests
This also more-or-less completes the transition to the "new" local CI
setup which is able to automatically count workspaces and features and
build feature matrices.

One issue is that this setup is extremely slow. It would almost
certainly be faster to fork crate2nix or something to build separate
derivations. Or at least allow amortizing the instantiations of the
dependency tree, or something.

Instantiating a test of a single rust-bitcoin commit takes 23 minutes.