local-ci: some shellcheck things
There are also shellcheck issues related to "don't do local x=$(y), use
two lines" which just seem annoying to fix. And a bunch related to
arrays and strings which I'm scared to fix because the existing code
works and I've had lots of bugs related to arrays and strings.