utils: drop old githubSrcs stuff
This old code is no longer used -- we instead compute commit lists
outside of the nix derivation and pass them inline as configuration.

This is for a couple of reasons. One is that Nix's git support is
atrocious, forcing us to IFD, which is always a slowdown. But worse than
that, even in an IFD, our choices for accessing the repo are "fetch a
single ref (which can't even be a bare commit, it must be a named ref)"
or "fetch every ref (every PR, every branch, etc)".

Obviously if we want to check multiple commits we have to choose the
latter ... meaning that to do a millisecond-long check that outputs a
couple hundred bytes' worth of git commit hashes, we need to import the
entire repo and all of its history to the nix store. This takes forever
and then slows down the nix-store filesystem because the result is that
pretty-much every check copies the current state of the git repo (10s of
000s of files on medium-sized repos) into the nix-store. Then
nix-optimize spends forever hardlinking all these objects together but
the filesystem still winds up with tens of millions of inodes and no fs
that I've tried can handle this in a reasonably speedy way. At least,
not on spinning disk drives.

There is some history here: https://github.com/NixOS/nix/issues/2431

The other reason, relatedly, is that Nix does not have any language
support for things like finding merge-bases, constructing commits, etc.
So we wind up needing to write all this code in bash anyway and at that
point might as well just call it outside Nix.

Given that the output of all this stuff is a list of commits, which are
either meaningless or deterministic depending whether we have the rest
of the repo ... we don't even gain any determinism from trying to use
Nix.

HAVING SAID ALL THIS, this commit doesn't actually fix the above
problems, only slightly improves them. We still have the srcFromCommit
function, which attempts to fetch a single commit by commit ID, and is
forced to set `allRefs = true` to do this. We want to eliminate this,
perhaps by creating a synthetic branch or something.