compute Cargo.nix on queue-pr and use database for lockfiles
This commit makes two main changes: one is that it pulls the lockfile
logic, which was previously split between queue-pr and run, into
queue-pr. (And there, it factors it out in the shellscript so that the
same logic is run for queue-pr, queue-commit, queue-merge.) This lets us
eliminate the "fallback lockfiles" in the Nix code as well as the
lockfile searching code in the Nix code.

Now, when you queue something, all the fallback logic happens there.

The second change is an abortive attempt to get rid of the Cargo.nix
IFD. At queue time, we run crate2nix and save the resulting .nix file in
the nix store. We associate it with the commit/locktime pair and pass it
into the nix code.

However, I wasn't able to get the nix code to make any use of this.
Committing here so that my working tree isn't too much of a mess, but I
hope to revisit this in the coming commits.