partially fix queue_merge
After changing the logic to use lockfiles from the database, we must
confront the fact that we compute merge commits "dynamically". That is,
when we queue a merge, we don't record the merge commit at the time of
queuing. We wait until we actually process the merge.

This is the correct behavior since we don't know the merge at the time
of queuing -- if we merge something in front of the newly-queued merge,
then our actual merge will differ from what it would've been.

Our existing logic cheats this by calling `$(git rev-parse HEAD)` from
inside the merge script and then dynamically trying to figure out what
lockfiles etc to use. We don't want to do that anymore because it
results in duplicated and hard to predict logic.

Instead, when queuing merges, we -will- record the merge commit. And
when we get to the actual merge, we'll compare trees to see if there is
any difference. If so, we'll immediately fail and requeue the merge.

In general, this will result in a quadratic number of queues (and texts
about requeing). But in practice this is probably manageable -- if we
have more than 5 or 6 merges queued up then (a) this will take all day
and (b) probably things will conflict and fail anyway. And 6 merges will
result in at most T(6) = 15 texts throughout the day.

This commit is a PARTIAL fix because it implements requeuing but doesn't
actually change which commit will be tried next time. So actually this
commit will just cause an infinite loop :). Will complete the fix
tomorrow.