local-ci: add "merge_pushes" table of commits that we intend to push
Rather than (ab)using the bitcoin-maintainer-tools merge script to run a test
prior to signing, we are going to replace it with one that does the same
checks and computes the same merge message, but which uses `jj` to
create the merge commit (without screwing around with any working
copies) and track it through its lifecycle of having a commit message
attached and being gpg-signed.

Then we are going to move the signing step outside of the main loop: we
request the user locate and sign commits, and we simply check whether
this has been done and push if so, updating other in-progress merges
accordingly and alerting the user if anything has changed.