sqlite: add standalone compute_sha512_root.py
We are migrating the github-merge.py logic into the local CI system. To
do this, we want access to the "SHA512 tree hash" used by github-merge
as a future-proof collision-resistant tree hash. Copy it into a
standalone Python file, put a small jj wrapper around it to allow use of
jj change IDs, and do CLI parsing.