local-ci.sh: require jj to be available on PATH
We are going to start using `jj describe` and other commands to track
and update merge commits. The `git` commands require checking out the
commits before manipulating them, which is untenable because there are
many situations where checkouts are disallowed or impossible or
destructive.

See https://wpsoftware.net/andrew/blog/moving-from-git-to-jujutsu.html