sqlite: add standalone compute_merge_description.py
This script does the merge commit message computation and sanity checks
from github-merge.py, and nothing else: it does not create merges, it
does not sign anything, etc.

On success this outputs the commit messsage on stdout and nothing else.
The idea is that this can be passed to `jj describe` if everything seems
good.