Attempt to get simplicity-webide working
The issue here is that the MSRV of simplicity-webide is 1.78 (I think) which
uses the new "version 4" lockfiles, which are identical to "version 3" except
that they've changed one character so that Rust 1.77- won't parse them.

However, Rust 1.77+ changed the output of cargo-metadata so that crate2nix
does not work with it.

No doubt this has been fixed upstream in crate2nix, but probably those updates
have broken stuff that I need..